pyproject.toml
[tool.poetry]
name = "pypi-token-client"
version = "0.1.0"
description = "Library and CLI tool for retrieving PyPI project tokens"
authors = ["smheidrich <smheidrich@weltenfunktion.de>"]
readme = "README.md"

[tool.poetry.scripts]
pypi-token-client = 'pypi_token_client:cli_main'

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.29"
keyring = "^23.13.1"
typer = "^0.7.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
black = "^22.10.0"
isort = "^5.10.1"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
